dom/workers/test/test_sharedWorker_lifetime.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
  <head>
    <title>Test for MessagePort and SharedWorkers</title>
    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  </head>
  <body>
    <script class="testbody" type="text/javascript">

var gced = false;
SpecialPowers.pushPrefEnv({ set: [["dom.workers.sharedWorkers.enabled", true]]},
function() {
  var sw = new SharedWorker('sharedWorker_lifetime.js');
  sw.port.onmessage = function(event) {
    ok(gced, "The SW is still alive also after GC");
    SimpleTest.finish();
  }

  sw = null;
  SpecialPowers.forceGC();
  gced = true;
});

SimpleTest.waitForExplicitFinish();
    </script>
  </body>
</html>
